src/main/java/com/cn/leetcode/A0112_hasPathSum.java
package com.cn.leetcode;

import com.cn.leetcode.internal.TreeNode;

/**
 * 112. 路径总和
 * 给你二叉树的根节点 root 和一个表示目标和的整数 targetSum ，判断该树中是否存在 根节点到叶子节点 的路径，这条路径上所有节点值相加等于目标和 targetSum 。
 * <p>
 * 叶子节点 是指没有子节点的节点。
 * <p>
 * <p>
 * <p>
 * 示例 1：
 * <p>
 * <p>
 * 输入：root = [5,4,8,11,null,13,4,7,2,null,null,null,1], targetSum = 22
 * 输出：true
 * 示例 2：
 * <p>
 * <p>
 * 输入：root = [1,2,3], targetSum = 5
 * 输出：false
 * 示例 3：
 * <p>
 * 输入：root = [1,2], targetSum = 0
 * 输出：false
 * <p>
 * <p>
 * 提示：
 * <p>
 * 树中节点的数目在范围 [0, 5000] 内
 * -1000 <= Node.val <= 1000
 * -1000 <= targetSum <= 1000
 */
public class A0112_hasPathSum {

    public boolean hasPathSum(TreeNode root, int targetSum) {
        if (root == null) return false;

        int d = targetSum - root.val;
        if (root.left == null && root.right == null) {
            return d == 0;
        }
        return hasPathSum(root.left, d) || hasPathSum(root.right, d);
    }
}
